Replaced merit gem with own code
We were assigning points to users in recalculatable ways, i.e instead of
adding points and keeping an inventory of those points, we may generate
those in realtime. That's why using custom logic. In future we may want to
use merit gem again if we're using levels and badges etc. Not until we're
not.